Adjust directory structure for users api
