+++ services/app-api/handlers/users/get/getUserById.js
@@ -1,5 +1,5 @@
-import handler from "./../../libs/handler-lib";
-import dynamoDb from "./../../libs/dynamodb-lib";
+import handler from "../../../libs/handler-lib";
+import dynamoDb from "../../../libs/dynamodb-lib";
 
 export const main = handler(async (event, context) => {
   // If this invokation is a prewarm, do nothing and return.
@@ -8,17 +8,20 @@ export const main = handler(async (event, context) => {
     return null;
   }
 
-  const data = JSON.parse(event.body);
-
   const params = {
     TableName:
       process.env.AUTH_USER_TABLE_NAME ?? process.env.AuthUserTableName,
     Key: {
-      userId: data.id,
+      userId: event.pathParameters["id"],
     },
   };
 
-  await dynamoDb.delete(params);
+  const result = await dynamoDb.get(params);
 
-  return { status: true };
+  if (!result.Item) {
+    return false;
+  }
+
+  // Return the retrieved item
+  return result.Item;
 });
