+++ services/app-api/handlers/users/post/deleteUser.js
@@ -1,5 +1,5 @@
-import handler from "./../../libs/handler-lib";
-import dynamoDb from "./../../libs/dynamodb-lib";
+import handler from "../../../libs/handler-lib";
+import dynamoDb from "../../../libs/dynamodb-lib";
 
 export const main = handler(async (event, context) => {
   // If this invokation is a prewarm, do nothing and return.
@@ -7,16 +7,18 @@ export const main = handler(async (event, context) => {
     console.log("Warmed up!");
     return null;
   }
+
+  const data = JSON.parse(event.body);
+
   const params = {
     TableName:
       process.env.AUTH_USER_TABLE_NAME ?? process.env.AuthUserTableName,
+    Key: {
+      userId: data.id,
+    },
   };
 
-  const result = await dynamoDb.scan(params);
+  await dynamoDb.delete(params);
 
-  if (!result.Items) {
-    throw new Error("No Users not found.");
-  }
-
-  return result.Items;
+  return { status: true };
 });
